refactor: 랭킹 리포지토리와 AWS 자격증명 구성을 정리
