--- src/main/java/com/LetMeDoWith/LetMeDoWith/domain/ranking/repository/RankingEntryRepository.java
@@ -1,10 +1,0 @@
-package com.LetMeDoWith.LetMeDoWith.domain.ranking.repository;
-
-import com.LetMeDoWith.LetMeDoWith.domain.ranking.model.RankingEntry;
-import java.util.List;
-
-public interface RankingEntryRepository {
-    List<RankingEntry> getEntriesByTopicAndRound(Long topicId, Long round);
-
-    List<RankingEntry> save(List<RankingEntry> rankingEntries);
-}
