--- src/main/java/com/LetMeDoWith/LetMeDoWith/domain/ranking/repository/RankingTopicRoundRepository.java
@@ -1,7 +1,0 @@
-package com.LetMeDoWith.LetMeDoWith.domain.ranking.repository;
-
-import com.LetMeDoWith.LetMeDoWith.domain.ranking.model.RankingTopicRound;
-
-public interface RankingTopicRoundRepository {
-    RankingTopicRound save(RankingTopicRound rankingTopicRound);
-}
