+++ src/main/java/com/LetMeDoWith/LetMeDoWith/infrastructure/config/AwsConfig.java
@@ -3,8 +3,9 @@ package com.LetMeDoWith.LetMeDoWith.infrastructure.config;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
 import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
-import software.amazon.awssdk.auth.credentials.DefaultCredentialsProvider;
+import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.presigner.S3Presigner;
@@ -15,9 +16,15 @@ public class AwsConfig {
     @Value("${cloud.aws.region}")
     private String region;
 
+    @Value("${cloud.aws.credentials.accessKey}")
+    private String accessKey;
+
+    @Value("${cloud.aws.credentials.secretKey}")
+    private String secretKey;
+
     @Bean
     public AwsCredentialsProvider awsCredentialsProvider() {
-        return DefaultCredentialsProvider.builder().profileName("letmedowith").build();
+        return StaticCredentialsProvider.create(AwsBasicCredentials.create(accessKey, secretKey));
     }
 
     @Bean
