+++ src/main/resources/application-dev-confidential.yml
@@ -28,6 +28,9 @@ auth:
 cloud:
   aws:
     region: ${AWS_REGION}
+    credentials:
+      accessKey: ${AWS_ACCESS_KEY_ID}
+      secretKey: ${AWS_SECRET_ACCESS_KEY}
     s3:
       bucketName: ${AWS_S3_BUCKET_NAME}
