+++ src/main/resources/application-local-confidential.yml
@@ -26,5 +26,8 @@ auth:
 cloud:
   aws:
     region: ${AWS_REGION}
+    credentials:
+      accessKey: ${AWS_ACCESS_KEY_ID}
+      secretKey: ${AWS_SECRET_ACCESS_KEY}
     s3:
       bucketName: ${AWS_S3_BUCKET_NAME}
